+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.0.1 under development
 
 - Enh #287: Allow `Dropdown::togglerVariant()` to be `null`, avoiding it setting a variant class (@Mister-42)
+- Bug #288: Add support for custom togglerClasses in `Dropdown` widget, and `addDropdownClass()` method (@terabytesoftw)
 
 ## 1.0.0 April 13, 2025